feat(TDC-7256/StepperForm) - Added new StepperForm component to centralize all the logic related to a Stepper flow
