+++ .changeset/silver-bulldogs-build.md
@@ -1,0 +1,5 @@
+---
+'@talend/react-stepper': minor
+---
+
+feat(TDC-7256/StepperForm) - Added new StepperForm component to centralize all the logic related to a Stepper flow
