+++ packages/stepper/custom.d.ts
@@ -1,3 +1,4 @@
+declare module '@talend/react-components';
 declare module '*.scss' {
 	const contents: Record<string, string>;
 	export default contents;
