+++ packages/stepper/src/components/StepperForm/StepFooter/index.ts
@@ -1,0 +1,3 @@
+import StepFooter from './StepFooter.component';
+
+export default StepFooter;
