+++ packages/stepper/src/components/StepperForm/StepHeader/StepHeader.types.ts
@@ -1,0 +1,6 @@
+export interface StepHeaderProps {
+	title: string;
+	subtitle?: string;
+}
+
+export type StepHeader = StepHeaderProps;
