--- packages/stepper/src/components/StepperForm/StepperFormFooter/index.ts
@@ -1,3 +1,0 @@
-import StepperFormFooter from './StepperFormFooter.component';
-
-export default StepperFormFooter;
