--- packages/stepper/src/components/StepperForm/StepperFormHeader/StepperFormHeader.types.ts
@@ -1,6 +1,0 @@
-export interface StepperFormHeaderProps {
-	title: string;
-	subtitle?: string;
-}
-
-export type StepperFormHeader = StepperFormHeaderProps;
