+++ packages/stepper/src/stories/Stepper.stories.tsx
@@ -1,12 +1,13 @@
 import { useState } from 'react';
-import StepperForm from '../src/components/StepperForm';
-import { StepperState } from '../src/hooks/useStepperForm/useStepperForm.types';
+import StepperForm from '../components/StepperForm';
+import { StepperState } from '../hooks/useStepperForm/useStepperForm.types';
 import {
 	BusinessContext,
 	FormComponentStep1,
 	FormComponentStep2,
 	FormComponentStep3,
 } from './Stepper.components';
+
 export default {
 	title: 'Stepper',
 	component: StepperForm,
@@ -23,42 +24,41 @@ export const DefaultStory = () => {
 	const initialSteps: StepperState = [
 		{
 			key: 'STEP1',
-			title: 'Step 1',
+			name: 'Step 1',
 			component: () => <FormComponentStep1 />,
+			header: {
+				title: 'This is step 1 title',
+			},
 		},
 		{
 			key: 'STEP2',
-			title: 'Step 2',
+			name: 'Step 2',
 			component: () => <FormComponentStep2 />,
+			header: {
+				title: 'This is step 2 title',
+			},
 		},
 		{
 			key: 'STEP3',
-			title: 'Step 3',
+			name: 'Step 3',
 			component: () => <FormComponentStep3 />,
+			header: {
+				title: 'This is step 3 title',
+			},
 		},
 	];
 	const [stepsData, setStepsData] = useState<any>({});
 
-	const BusinessComponent = () => {
-		return (
-			<StepperForm
-				header={{
-					title: 'This is step 1 title',
-				}}
-			/>
-		);
-	};
-
 	return (
-		<StepperForm.ContextProvider initialStepIndex={0} steps={initialSteps}>
+		<StepperForm.Provider initialStepIndex={0} steps={initialSteps}>
 			<BusinessContext.Provider
 				value={{
 					setStepsData,
 					stepsData,
 				}}
 			>
-				<BusinessComponent />
+				<StepperForm />
 			</BusinessContext.Provider>
-		</StepperForm.ContextProvider>
+		</StepperForm.Provider>
 	);
 };
